% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{defaultStrategies}
\alias{defaultStrategies}
\title{Strategies Provided with the \code{axelRod} Package}
\usage{
defaultStrategies()
}
\value{
A list of lists. Each sublist will have three elements:
 \describe{
   \item{name}{A character string giving the name of the strategy.}
   \item{description}{A character string describing the strategy.}
   \item{fn}{The function implementing the strategy.}
 }
}
\description{
This function returns a list of strategies that can be used to
 run an Axelrod-like game theory tournament.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

