% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tournament.R
\name{shiny_tournament}
\alias{shiny_tournament}
\title{Axelrod Tournament Shiny App}
\usage{
shiny_tournament(payoff = matrix(c(3, 5, 0, 1), nrow = 2),
  custom_strategies = NULL, with_default_strategies = TRUE)
}
\arguments{
\item{payoff}{A 2x2 payoff matrix of the tournament.}

\item{custom_strategies}{A list of custom strategies, following the model of
the list returned by \code{\link{defaultStrategies}}.}

\item{with_default_strategies}{If TRUE, the tournament app uses the default
strategies returned by \code{\link{defaultStrategies}} (the default), in
addition of any custom strategy provided by the user. If FALSE, it uses only
the custom strategies provided by the user.}
}
\value{
This function launches a Shiny application.
}
\description{
This function launches a Shiny app to run "Prisoner's Dilemma"
 matches between pairs of strategies.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

